Make treasure a type of Item

+++ DungeonCrawl/Game/Systems/TurnTakingSystem.swift
@@ -123,12 +123,15 @@ class TurnTakingSystem: System, TurnTaking {
     }
     
     private func pickUpItems(at cell: GridCell, for actor: Entity, inventory: InventoryComponent) {
-        for item in entityManager.entities(with: TreasureComponent.self) {
-            guard let treasure = item.treasureComponent() else { fatalError("Entity should have treasure component") }
+        for item in entityManager.entities(with: ItemComponent.self) {
+            guard let itemComponent = item.itemComponent() else { fatalError("Entity should have treasure component") }
             guard let itemSprite = item.spriteComponent(), itemSprite.cell == cell else { continue }
-            inventory.gold += treasure.gold
-            showPickedUpTreasureMessage(actor: actor, treasureDescription: "\(treasure.gold) gold pieces")
-            entityManager.remove(entity: item)
+            if itemComponent.item.isTreasure {
+                let gold = itemComponent.item.value
+                inventory.gold += gold
+                showPickedUpTreasureMessage(actor: actor, treasureDescription: "\(gold) gold pieces")
+                entityManager.remove(entity: item)
+            }
         }
     }
